requirements.txt
pyautogen>=0.2.19
ipython
huggingface_hub
matplotlib

# financial libraries
finnhub-python
yfinance
mplfinance

# data handling
numpy
pandas
pyPDF2

# packaging
setuptools>=41.4.0
wheel>=0.33.6

# historical data
tushare

pandas_datareader